Remove unwated code lines and Insert user request statuses

+++ core/database/migrations/2022_08_31_170057_create_user_request_statuses_table.php
@@ -14,7 +14,7 @@ class CreateUserRequestStatusesTable extends Migration
     public function up()
     {
         Schema::create('user_request_statuses', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->integer('id');
             $table->string('name');
             $table->timestamps();
         });
